Implementing (WIP) the GCI-GOI YAGA algorithm

+++ inc/gci_sedreams.h
@@ -2,6 +2,7 @@
 #define GCI_SEDREAMS_H
 
 #include <deque>
+#include <utility>
 #include <vector>
 
 #include <gsl/gsl_vector.h>
@@ -10,6 +11,6 @@ std::vector<size_t> gci_sedreams(gsl_vector *signal, const double fs, const doub
 
 std::deque<size_t> findPeaks(gsl_vector *signal, double sign);
 
-int median(std::vector<int>& x);
+double median(std::vector<double>& x);
 
 #endif // GCI_SEDREAMS_H
